"rescue" keyword without argument only catches subclasses of StandardError
In order to catch errors that don't inherit from StandardError, you need to 
'rescue Exception', otherwise the exception will just blow past and bubble up.

In addition, when rescuing an exception and raising a different one, it is 
useful to add the "caller" as a third argument to raise, which will preserve
the stack trace from the original exception, aiding debugging the actual cause
of the error.