In test_net_http_raises_continual_eof_exceptions server does not respond_to?(:address)
I'm not sure if this is a bug in the tests, or if this will happen in real life, but
server in this context does not always respond_to?(:address) and this causes a 
NoMethodError.

This commit uses server.address if server.responds_to?(:address), otherwise assumes
server is a String, and uses it directly in creating the exception message.

Additionally, preserves the stack when raising CloudFiles::Exception::Connection