Consolidate storage and cdn requests into specific methods
Instead of constructing cdn or storage paths in each and every 
method on the connection object, generalize the concept of 
"making a {storage,cdn} request" into methods for each. This allows
us to simplify making a request in other methods, and centralizes
the generation of request paths, which in turn (will eventually) make 
mocking and testing simpler.