Use Connection#cdn_request and Connection#storage_request in Container and StorageObject
Reduce the duplication in Container and StorageObject by using the
storage_request and cdn_request methods on Connection. Currently, the complete
cfreq call must be constructed in both the Container and StorageObject for each
and every request being made using data duplicated from the Connection on
initialization. By using the *_request methods on the related Connection
instance, we no longer need to copy this information into instance variables in
these children of a given Connection. 

This is an improvement in "information hiding", as only the Connection needs to
track this information (simplifying maintenance going forward if the format of
these requests changes in the future). This also simplifies mocking in the tests.