Clean up return value and metadata state management for set_*_acl methods
This commit fixes two small issues with the set_read_acl and set_write_acl
methods.

Currently the set_*_acl methods don't reset the cached metadata, so after
changing the value server-side, the read_acl and write_acl methods will still
return the old values. This patch fixes that behavior so that the new values
will be fetched from the server on the next invocation.

The return value currently is the response code of the HTTP request sent to the
server, which while not erroneous (as it would be a non-nil or false value), is
not consistent with the rest of the API. 

This commit also adds tests for error cases and return values.